feat: don't download fonts if already present

+++ install/fonts.sh
@@ -4,6 +4,9 @@
 
 set -euo pipefail
 
+# download Meslo Nerd Font patched for Powerlevel10k
+# see https://github.com/romkatv/powerlevel10k#meslo-nerd-font-patched-for-powerlevel10k
+#
 # contents of this file were originally extracted from
 # https://github.com/romkatv/powerlevel10k/blob/5ee7847/internal/wizard.zsh
 
@@ -37,9 +40,11 @@ function install_font() {
       local style
       for style in Regular Bold Italic 'Bold Italic'; do
         local file="MesloLGS NF ${style}.ttf"
-        run_command "Downloading %B$file%b" \
-          curl -fsSL -o ~/Library/Fonts/$file.tmp "$font_base_url/${file// /%20}"
-        command mv -f -- ~/Library/Fonts/$file{.tmp,} || quit -c
+        if [[ ! -f ~/Library/Fonts/$file ]]; then
+            run_command "Downloading %B$file%b" \
+            curl -fsSL -o ~/Library/Fonts/$file.tmp "$font_base_url/${file// /%20}"
+            command mv -f -- ~/Library/Fonts/$file{.tmp,} || quit -c
+        fi
       done
 }
